requirements.txt
# Production Django Backend Requirements
# Python 3.12+ required (tested with Python 3.12.x)

# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter
dj-database-url==2.1.0

# Async support
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0  # ASGI server
uvicorn[standard]==0.27.0  # ASGI server with workers
gunicorn==21.2.0  # WSGI server

# Supabase client
# Using 1.2.0 which is more stable and doesn't have the proxy argument issue
supabase==1.2.0
postgrest==0.11.0  # Compatible with supabase 1.2.0 (requires <0.12.0 and >=0.10.8)

# AI Providers
openai==1.12.0
anthropic==0.18.1
assemblyai==0.28.0

# Twilio
twilio==9.1.0

# Google Cloud
google-cloud-tasks==2.16.1
google-cloud-run==0.10.0
google-auth==2.27.0

# Utilities
python-dotenv==1.0.0
pydantic==2.6.1
python-multipart==0.0.6
httpx==0.24.1  # Compatible with supabase 1.2.2
aiofiles==23.2.1  # Async file operations

# Logging & Monitoring
structlog==24.1.0
python-json-logger==2.0.7

# Security
django-ratelimit==4.1.0
cryptography==42.0.2

# Testing
pytest==7.4.4
pytest-django==4.8.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
factory-boy==3.3.0

# Development
black==24.1.1
ruff==0.1.11
mypy==1.8.0
django-stubs==4.2.7

